refact: 전역 api 파일 분리
